requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Monitoring & Logging
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Security & Rate Limiting
django-ratelimit==4.1.0

# Utilities
pytz==2023.3

# WSGI Server
gunicorn==21.2.0

# Webhook signature validation
cryptography==41.0.7
